Update Parts_1&2&3
Added 4(b). The main issue was the time of execution. The problem is massive in terms of calculation and I did not find the way to optimise it. Basically it estimates the chi-squared value 10000 times which is wicked. Although the result is promising, it took about 3 hours to compile. Hence I decided to save the results into a separate file 'Chi_Set_background' so each time you run the code the data is uploaded without performing any calculations.